pyproject.toml
[project]
name = "facial-keypoints-detection"
version = "0.1.0"
description = "Facial keypoints detection experiment pipeline"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "numpy>=1.26,<2.3",
    "pandas>=2.1",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "torch>=2.2",
    "torchvision>=0.17",
    "pillow>=10.2",
    "pyyaml>=6.0.3",
]

[tool.uv]
package = false
